See http://fyireporting.com/forum/viewtopic.php?t=1068&postdays=0&postorder=asc&start=15
The Lexer.cs file to allow for shortcuts for the Fields, Parameters, and Globals.

Fields can now be entered as {FieldName} instead of Fields!FieldName.Value,
Parameters can be entered as {?ParameterName} instead of Parameters!ParameterName.Value,
User fields can be entered as {!UserField} instead of
User!UserField,
and Globals can be entered as {@GlobalName} instead of Globals!GlobalName.