fix chart drawing, add conditional compilation for image handling
On windows it would use _mf metafile but would save and write _bm for Bitmap.
It resulted in a black rectangle instead of charts.   This bug was introduced earlier
while working on the drawing compat.

Introduce conditional compilation using the `DRAWINGCOMPAT` symbol.
In the `Save` method, modify the behavior to save `_mf` if it is
not null and `DRAWINGCOMPAT` is not defined; otherwise, save `_bm`.
In the `Image` method, return `_mf` if it is not null and
`DRAWINGCOMPAT` is not defined; otherwise, return `_bm`.
This change allows more flexibility in handling different image
formats based on the `DRAWINGCOMPAT` symbol.