Remove net48 ItemGroup with System.Web.Services reference
The changes in the `Majorsilence.Reporting.DataProviders.csproj` file involve the removal of an `ItemGroup` that was conditionally included for the `net48` target framework. This `ItemGroup` contained a reference to `System.Web.Services`. The rest of the project file remains unchanged. This means that the project will no longer include the `System.Web.Services` reference when targeting the .NET Framework 4.8.