add PDF rendering tests for multiple barcode types
Tests currently fail due to a bug in the the reporting
library.   The correct barcode types are ignored when set via parameters and rendering to files.

A future commit will fix the bug.   The intent of this
commit to to ensure once the bug is fixed it is not
reintroduced at a later time.